+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Make the claim→principal mapping configurable.
+  There is no longer any default mapper so no principals will be set unless you set one in the config.
 
 ## [0.1.9] - 2024-09-18
 ### Fixed